pyproject.toml
[project]
name = "langgraph-explained"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "langchainhub>=0.1.21",
    "langgraph>=0.6.4",
    "openai>=1.99.3",
    "pinecone>=7.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
